pyproject.toml
[tool.poetry]
name = "2ximages"
version = "0.1.0"
description = ""
authors = ["TooriG <gakumotoori@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
